Move core.tests for api_view to new file in utils.tests
